requirements.txt
python-telegram-bot==20.3
pypdf==3.17.4
python-docx==1.1.0
reportlab==4.0.8
psutil==5.9.5
pdf2image==1.16.3
Pillow>=10.0.0,<11.0.0
Flask==3.0.0
requests==2.31.0
